move the tooltip range to base char
