+++ src/sentry/static/sentry/app/components/charts/components/tooltip.tsx
@@ -1,5 +1,6 @@
 import 'echarts/lib/component/tooltip';
 import {EChartOption} from 'echarts';
+import moment from 'moment';
 
 import {getFormattedDate, getTimeFormat} from 'app/utils/dates';
 import BaseChart from 'app/components/charts/baseChart';
@@ -12,14 +13,35 @@ function defaultFormatAxisLabel(
   value: number,
   isTimestamp: boolean,
   utc: boolean,
-  showTimeInTooltip: boolean
+  showTimeInTooltip: boolean,
+  bucketSize?: number
 ) {
   if (!isTimestamp) {
     return value;
   }
-  const format = `MMM D, YYYY ${showTimeInTooltip ? getTimeFormat() : ''}`.trim();
 
-  return getFormattedDate(value, format, {local: !utc});
+  if (!bucketSize) {
+    const format = `MMM D, YYYY ${showTimeInTooltip ? getTimeFormat() : ''}`.trim();
+    return getFormattedDate(value, format, {local: !utc});
+  }
+
+  const now = moment();
+  const bucketStart = moment(value);
+  const bucketEnd = moment(value + bucketSize);
+
+  const showYear = now.year() !== bucketStart.year() || now.year() !== bucketEnd.year();
+  const showEndDate = bucketStart.date() !== bucketEnd.date();
+
+  const formatStart = `MMM D${showYear ? ', YYYY' : ''} ${
+    showTimeInTooltip ? getTimeFormat() : ''
+  }`.trim();
+  const formatEnd = `${showEndDate ? `MMM D${showYear ? ', YYYY' : ''} ` : ''}${
+    showTimeInTooltip ? getTimeFormat() : ''
+  }`.trim();
+
+  return `${getFormattedDate(bucketStart, formatStart, {
+    local: !utc,
+  })} - ${getFormattedDate(bucketEnd, formatEnd, {local: !utc})}`;
 }
 
 function defaultValueFormatter(value: string | number) {
@@ -48,7 +70,7 @@ function getSeriesValue(series: EChartOption.Tooltip.Format, offset: number) {
   return undefined;
 }
 
-type NeededChartProps = 'isGroupedByDate' | 'showTimeInTooltip' | 'utc';
+type NeededChartProps = 'isGroupedByDate' | 'showTimeInTooltip' | 'utc' | 'bucketSize';
 
 type TooltipFormatters =
   | 'truncate'
@@ -67,6 +89,7 @@ function getFormatter({
   truncate,
   formatAxisLabel,
   utc,
+  bucketSize,
   valueFormatter = defaultValueFormatter,
   nameFormatter = defaultNameFormatter,
 }: FormatterOptions) {
@@ -103,7 +126,8 @@ function getFormatter({
         timestamp,
         !!isGroupedByDate,
         !!utc,
-        !!showTimeInTooltip
+        !!showTimeInTooltip,
+        bucketSize
       );
       // eCharts sets seriesName as null when `componentType` !== 'series'
       const truncatedName = truncationFormatter(
@@ -140,7 +164,13 @@ function getFormatter({
 
     const label =
       seriesParams.length &&
-      axisFormatterOrDefault(timestamp, !!isGroupedByDate, !!utc, !!showTimeInTooltip);
+      axisFormatterOrDefault(
+        timestamp,
+        !!isGroupedByDate,
+        !!utc,
+        !!showTimeInTooltip,
+        bucketSize
+      );
 
     return [
       '<div class="tooltip-series">',
@@ -172,6 +202,7 @@ export default function Tooltip({
   formatter,
   truncate,
   utc,
+  bucketSize,
   formatAxisLabel,
   valueFormatter,
   nameFormatter,
@@ -185,6 +216,7 @@ export default function Tooltip({
       showTimeInTooltip,
       truncate,
       utc,
+      bucketSize,
       formatAxisLabel,
       valueFormatter,
       nameFormatter,
